Refactored component into ui, helpers, types, etc.
