+++ src/features/DropzoneCSV/lib/types.ts
@@ -1,0 +1,9 @@
+export enum CSVSeparator {
+    SEMICOLON = 'semicolon',
+    COMMA = 'comma'
+}
+
+export interface IRegexParser {
+    regValid: RegExp,
+    regValue: RegExp,
+}
