+++ src/features/DropzoneCSV/ui/DropzoneCSV.tsx
@@ -1,0 +1,92 @@
+import React, {useRef} from 'react';
+import UploadSvg from '../../../shared/images/file-upload.svg'
+import cls from './Dropzone.module.css'
+import {ab2str, CSVtoArray} from "../lib/helpers";
+
+export const DropzoneCsv = () => {
+    const [isDragActive, setIsDragActive] = React.useState(false);
+    const inputRef = useRef<HTMLInputElement>(null);
+
+    // handle drag events
+    const handleDrag = function(e: React.DragEvent<HTMLDivElement | HTMLFormElement>) {
+        e.preventDefault();
+        e.stopPropagation();
+        if (e.type === "dragenter" || e.type === "dragover") {
+            setIsDragActive(true);
+        } else if (e.type === "dragleave") {
+            setIsDragActive(false);
+        }
+    };
+
+    // triggers when file is dropped
+    const handleDrop = function(e: React.DragEvent<HTMLDivElement>) {
+        e.preventDefault();
+        e.stopPropagation();
+        setIsDragActive(false);
+        if (e.dataTransfer.files && e.dataTransfer.files[0]) {
+            handleFile(e.dataTransfer.files[0]);
+        }
+    };
+
+    const handleFile = (file: File) => {
+        let fr = new FileReader();
+        fr.onload = receivedText;
+        fr.readAsText(file);
+
+        function receivedText() {
+            let result:string
+            if (fr.result) {
+                // convert to string
+                if(typeof fr.result !== 'string') {
+                    result = ab2str(fr.result)
+                } else {
+                    result = fr.result
+                }
+                // turn string to 2d array
+                let data = CSVtoArray(result)
+                // data ready to send to backend
+            }
+        }
+    }
+
+    // triggers when file is selected
+    const handleChange = function(e: React.ChangeEvent<HTMLInputElement>) {
+        e.preventDefault();
+        if (e.target.files && e.target.files[0]) {
+            handleFile(e.target.files[0]);
+        }
+    };
+
+    // triggers the input when the button is clicked
+    const onButtonClick = () => {
+        inputRef.current?.click();
+    };
+
+    return (
+        <form
+            className={cls.form}
+            onDragEnter={handleDrag}
+            onSubmit={(e) => e.preventDefault()}
+        >
+            <input
+                id="input-file-upload"
+                ref={inputRef} type="file"
+                className={`${cls.input}`}
+                multiple={false}
+                onChange={handleChange}
+            />
+            <label className={`${cls.label} ${isDragActive ? cls.dragActive : ""}`} htmlFor="input-file-upload">
+                <UploadSvg/>
+                <button className={cls.uploadButton} onClick={onButtonClick}>Загрузите файл</button>
+            </label>
+            { isDragActive && <div
+                className={cls.dragFile}
+                onDragEnter={handleDrag}
+                onDragLeave={handleDrag}
+                onDragOver={handleDrag}
+                onDrop={handleDrop}
+            />}
+        </form>
+    );
+};
+
